comments cleanup
