+++ src/App.css
@@ -1,7 +1,1 @@
-/* @import url("http://fonts.cdnfonts.com/css/iranian-sans"); */
 @import url("http://fonts.cdnfonts.com/css/iransansx");
-body {
-  /* font-family: "Iranian Sans", sans-serif; */
-  /* font-family: "IRANSansX", sans-serif; */
-  background-color: "#f5f6fb";
-}
