--- src/components/ScrollableTabsButtonPrevent.js
@@ -1,93 +1,0 @@
-import * as React from "react";
-import PropTypes from "prop-types";
-import SwipeableViews from "react-swipeable-views";
-import { useTheme } from "@mui/material/styles";
-import AppBar from "@mui/material/AppBar";
-import Tabs from "@mui/material/Tabs";
-import Tab from "@mui/material/Tab";
-import Typography from "@mui/material/Typography";
-import Box from "@mui/material/Box";
-
-export default function ScrollableTabsButtonPrevent() {
-  const [value, setValue] = React.useState(0);
-  const theme = useTheme();
-  const handleChange = (event, newValue) => {
-    setValue(newValue);
-  };
-  const handleChangeIndex = (index) => {
-    setValue(index);
-  };
-
-  function TabPanel(props) {
-    const { children, value, index, ...other } = props;
-
-    return (
-      <div
-        role="tabpanel"
-        hidden={value !== index}
-        id={`full-width-tabpanel-${index}`}
-        aria-labelledby={`full-width-tab-${index}`}
-        {...other}
-      >
-        {value === index && (
-          <Box sx={{ p: 3 }}>
-            <Typography>{children}</Typography>
-          </Box>
-        )}
-      </div>
-    );
-  }
-  // console.log(theme.direction);
-
-  return (
-    <Box sx={{ bgcolor: "background.paper" }}>
-      {/* <Box sx={{ maxWidth: { xs: 320, sm: 480 }, bgcolor: "background.paper" }}> */}
-      <Tabs
-        value={value}
-        onChange={handleChange}
-        variant="scrollable"
-        scrollButtons={false}
-        aria-label="scrollable prevent tabs example"
-      >
-        <Tab label="Item One" />
-        <Tab label="Item Two" />
-        <Tab label="Item Three" />
-        <Tab label="Item Four" />
-        <Tab label="Item Five" />
-        <Tab label="Item Six" />
-        <Tab label="Item Seven" />
-      </Tabs>
-
-      <SwipeableViews
-        axis={theme.direction === "rtl" ? "x-reverse" : "x"}
-        index={value}
-        onChangeIndex={handleChangeIndex}
-      >
-        <TabPanel value={value} index={0} dir={theme.direction}>
-          Item One
-        </TabPanel>
-        <TabPanel value={value} index={1} dir={theme.direction}>
-          Item Two
-        </TabPanel>
-        <TabPanel value={value} index={2} dir={theme.direction}>
-          Item Three
-        </TabPanel>
-        <TabPanel value={value} index={3} dir={theme.direction}>
-          Item Four
-        </TabPanel>
-        <TabPanel value={value} index={4} dir={theme.direction}>
-          Item five
-        </TabPanel>
-        <TabPanel value={value} index={5} dir={theme.direction}>
-          Item Three
-        </TabPanel>
-        <TabPanel value={value} index={6} dir={theme.direction}>
-          Item six
-        </TabPanel>
-        <TabPanel value={value} index={7} dir={theme.direction}>
-          Item seven
-        </TabPanel>
-      </SwipeableViews>
-    </Box>
-  );
-}
