+++ src/constants/languageConvertor.js
@@ -1,42 +1,3 @@
-// const finglish = {
-//   "Esteghlal Molasani",
-//   "Arman Gohar",
-//   "Pars Janoobi Jam",
-//   "Astara",
-//   "Kheybar Khorramabad",
-//   "Mes Kerman",
-//   "Khooshe Talaee",
-//   "Qashqai Shiraz",
-//   "Malavan",
-//   "Vista Toorbin",
-//   "Mes Shahr Babak",
-//   "Baderan Tehran",
-//   "Machine Sazi FC",
-//   "Shahin Bushehr",
-//   "Saipa",
-//   "Khooshe Talaee Saveh",
-//   "Rayka Babol",
-//   "Shahrdari Astara",
-//   "Shahrdari Hamedan",
-//   "Shams Azar Qazvin",
-//   "Aluminium Arak",
-//   "Mes Rafsanjan",
-//   "Naft Masjed Soleyman",
-//   "Fajr Sepasi",
-//   "Gol Gohar",
-//   "Nassaji Mazandaran",
-//   "Esteghlal FC",
-//   "Foolad FC",
-//   "Tractor Sazi",
-//   "ZOB Ahan",
-//   "Padideh Khorasan",
-//   "Paykan",
-//   "Persepolis FC",
-//   "Sanat Naft" ,
-//   "Sepahan FC" ,
-//   "Havadar",
-// };
-
 export const translate = {
   "Esteghlal Mollasani": "استقلال ملاسانی",
   "Arman Gohar": "آرمان گوهر",
